+++ src/Symfony/Component/DependencyInjection/Tests/Compiler/IntegrationTest.php
@@ -20,7 +20,6 @@ use Symfony\Component\DependencyInjection\ChildDefinition;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
-use Symfony\Component\DependencyInjection\Exception\LogicException;
 use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
 use Symfony\Component\DependencyInjection\Reference;
 use Symfony\Component\DependencyInjection\ServiceLocator;
@@ -33,7 +32,6 @@ use Symfony\Component\DependencyInjection\Tests\Fixtures\IteratorConsumer;
 use Symfony\Component\DependencyInjection\Tests\Fixtures\LocatorConsumer;
 use Symfony\Component\DependencyInjection\Tests\Fixtures\LocatorConsumerConsumer;
 use Symfony\Component\DependencyInjection\Tests\Fixtures\LocatorConsumerFactory;
-use Symfony\Component\DependencyInjection\Tests\Fixtures\MultipleArgumentBindings;
 use Symfony\Component\DependencyInjection\Tests\Fixtures\TaggedService1;
 use Symfony\Component\DependencyInjection\Tests\Fixtures\TaggedService2;
 use Symfony\Component\DependencyInjection\Tests\Fixtures\TaggedService3;
@@ -338,7 +336,7 @@ class IntegrationTest extends TestCase
             ->addTag('foo_bar', ['foo' => 'foo'])
         ;
         $container->register(IteratorConsumer::class)
-            ->setAutoconfigured(true)
+            ->setAutowired(true)
             ->setPublic(true)
         ;
 
@@ -391,7 +389,7 @@ class IntegrationTest extends TestCase
             ->addTag('foo_bar', ['foo' => 'foo'])
         ;
         $container->register(LocatorConsumer::class)
-            ->setAutoconfigured(true)
+            ->setAutowired(true)
             ->setPublic(true)
         ;
 
@@ -419,7 +417,7 @@ class IntegrationTest extends TestCase
             ->setPublic(true)
             ->setArguments([
                 (new Definition(LocatorConsumer::class))
-                    ->setAutoconfigured(true),
+                    ->setAutowired(true),
             ])
         ;
 
@@ -445,7 +443,7 @@ class IntegrationTest extends TestCase
         $container->register(LocatorConsumerFactory::class);
         $container->register(LocatorConsumer::class)
             ->setPublic(true)
-            ->setAutoconfigured(true)
+            ->setAutowired(true)
             ->setFactory(new Reference(LocatorConsumerFactory::class))
         ;
 
@@ -456,22 +454,6 @@ class IntegrationTest extends TestCase
 
         $locator = $s->getLocator();
         self::assertSame($container->get(FooTagClass::class), $locator->get('my_service'));
-    }
-
-    /**
-     * @requires PHP 8
-     */
-    public function testMultipleArgumentBindings()
-    {
-        $container = new ContainerBuilder();
-        $container->register(MultipleArgumentBindings::class)
-            ->setPublic(true)
-            ->setAutoconfigured(true)
-        ;
-
-        $this->expectException(LogicException::class);
-        $this->expectExceptionMessage('Cannot autoconfigure argument "$collection": More than one autoconfigurable attribute found.');
-        $container->compile();
     }
 
     public function testTaggedServiceWithDefaultPriorityMethod()