+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/MultipleArgumentBindings.php
@@ -2,13 +2,13 @@
 
 namespace Symfony\Component\DependencyInjection\Tests\Fixtures;
 
-use Symfony\Component\DependencyInjection\Attribute\BindTaggedIterator;
-use Symfony\Component\DependencyInjection\Attribute\BindTaggedLocator;
+use Symfony\Component\DependencyInjection\Attribute\TaggedIterator;
+use Symfony\Component\DependencyInjection\Attribute\TaggedLocator;
 
 final class MultipleArgumentBindings
 {
     public function __construct(
-        #[BindTaggedIterator('my_tag'), BindTaggedLocator('another_tag')]
+        #[TaggedIterator('my_tag'), TaggedLocator('another_tag')]
         object $collection
     ) {
     }